refactor: move EventModal from Calendar to page for better architecture
